+++ tests/dummy/app/controllers/index.js
@@ -1,9 +1,13 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
+  isShort: true,
   actions: {
     log(message) {
       console.log(message);
+    },
+    toggleHeight() {
+      this.toggleProperty('isShort');
     }
   }
 });