+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+## 0.3.4
+
+### Pull Requests
+
+- [#25](https://github.com/alphasights/ember-scrollable/pull/25)  scrollTo and onScroll *by [Jan Buschtöns](https://github.com/buschtoens)*
+
+## 0.3.3
+
+### Pull Requests
+
+- [#17](https://github.com/alphasights/ember-scrollable/pull/17)  Allow addon to be nested at depth of n  *by [Offir Golan](https://github.com/offirgolan)*
+- [#16](https://github.com/alphasights/ember-scrollable/pull/16)  Add scrollTop action which can be triggered from outside  *by [crudo1f/feature](https://github.com/crudo1f/feature)*
+- [#14](https://github.com/alphasights/ember-scrollable/pull/14)  Bail from createScrollbar if isDestroyed.  *by [blimmer/bug](https://github.com/blimmer/bug)*
+- [#21](https://github.com/alphasights/ember-scrollable/pull/21)  Fix horizontal scroll demo  *by [Kirill Shaplyko](https://github.com/Baltazore)*
+- [#20](https://github.com/alphasights/ember-scrollable/pull/20)  Updated to Ember 2.8  *by [Kirill Shaplyko](https://github.com/Baltazore)*
+- [#23](https://github.com/alphasights/ember-scrollable/pull/23)  Removed hide delay in tests *by [Taras Mankovski](https://github.com/taras)*
+
 ## 0.3.2
 
 ## 0.3.1